Add support for scalars and numpy in torch.testing
Add support to compare scalars as well as `np.ndarray`'s with torch.testing. We are reusing the mathcing functionality that is already in place for tensors, by casting the inputs. The approach can easily extended if we want to support other input types as long as they can be cast to a tensor.

ghstack-source-id: e7e998dc94ff795ddbe168f4fecb909e4651040c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55786